change how the code is run
the apl file can be run as a script with dyalog, instead of outputing
the contents of the file directly.  this also doesn't show the initial
license message.

to make this work, the indices of the command line arguments also had to
be shifted by two, since the second and third arguments are now '-script
/src/frb.apl'.